+++ server/database/entity/content.go
@@ -15,6 +15,18 @@ const (
 	SHOW_EPISODE ContentType = "tv_episode"
 )
 
+// NOTE: Can return empty string, for ContentTypes that can't map to
+// SupportedMedia.
+func ContentTypeToSupportedMedia(ct ContentType) util.SupportedMedia {
+	switch ct {
+	case MOVIE:
+		return util.SupportedMediaMovie
+	case SHOW:
+		return util.SupportedMediaShow
+	}
+	return ""
+}
+
 // For storing cached content, so we can serve the basic local data for watched list to work
 type Content struct {
 	ID               int         `json:"id" gorm:"primaryKey;autoIncrement"`
@@ -42,11 +54,5 @@ func (c *Content) GetTypeSupportedMedia() util.SupportedMedia {
 	if c == nil {
 		return ""
 	}
-	switch c.Type {
-	case MOVIE:
-		return util.SupportedMediaMovie
-	case SHOW:
-		return util.SupportedMediaShow
-	}
-	return ""
+	return ContentTypeToSupportedMedia(c.Type)
 }